feat: Add `/tmp` as a volume
Reason behind that is that /tmp is often used by programs and our
pipelines to write and read temporary files in /tmp. Since /tmp is on
the overlay filesystem created by docker it will be slower then a volume
mount. So adding it as a volume should slightly improve pipeline
performance.